requirements.txt
asttokens==2.2.1
backcall==0.2.0
certifi==2023.5.7
charset-normalizer==3.1.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.1.3
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
executing==1.2.0
filelock==3.12.0
flatbuffers==23.5.26
fonttools==4.39.4
humanfriendly==10.0
idna==3.4
ipykernel==6.23.1
ipython==8.13.2
jedi==0.18.2
Jinja2==3.1.2
jupyter_client==8.2.0
jupyter_core==5.3.0
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mpmath==1.3.0
nest-asyncio==1.5.6
networkx==3.1
numpy==1.24.3
onnx==1.14.0
onnxruntime==1.15.0
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
packaging==23.1
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.1
prompt-toolkit==3.0.38
protobuf==4.23.2
psutil==5.9.5
pure-eval==0.2.2
Pygments==2.15.1
pyparsing==3.0.9
pyreadline3==3.4.1
python-dateutil==2.8.2
pywin32==306
pyzmq==25.1.0
requests==2.31.0
segment-anything==1.0
six==1.16.0
stack-data==0.6.2
sympy==1.12
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tornado==6.3.2
traitlets==5.9.0
typing_extensions==4.6.1
urllib3==2.0.2
wcwidth==0.2.6
wget==3.2
uvicorn==8.1.2
fastapi==8.1.2

